requirements.txt
flask==3.0.0
openai==1.3.0
httpx==0.25.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
scikit-learn==1.3.2
numpy==1.24.3
pyaudio==0.2.13